+++ FakeServer/Common/Config.cs
@@ -5,5 +5,6 @@
         public const string ApiRoute = "api";
         public const string AsyncRoute = "async";
         public const string GraphQLRoute = "graphql";
+        public const string TokenRoute = "token";
     }
 }